tests/wpt/web-platform-tests/old-tests/submission/Microsoft/sandbox/sandbox_002.htm

<!DOCTYPE html>
<html>
<head>
    <title>HTML5 Sandbox: Allow autoplay for HTML5 Video inside iframe with sandbox attribute if sandbox='allow-scripts'.</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type" />
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://dev.w3.org/html5/spec/Overview.html#sandboxed-automatic-features-browsing-context-flag" />
    <meta name="assert" content="Allow autoplay for HTML5 Video inside iframe with sandbox attribute if sandbox='allow-scripts'." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script type="text/javascript">
    
        
        var t = async_test("Allow autoplay for HTML5 Video inside iframe with sandbox attribute if sandbox='allow-scripts'.");
  
        function callback(event)
        {
            t.step(function(){ 
                assert_true('sandbox' in document.createElement('iframe')); 
                assert_equals(event.data, "play event fired");
            });
            t.done();
        }

        var timer = setTimeout(callback, 8000);
        window.addEventListener("message", callback, false);
    </script>
    <iframe src="iframe_sandbox_002.htm" sandbox="allow-scripts" style="display: none"></iframe>
    <div id=log></div>
</body>
</html>
